requirements.txt
langchain>=0.1.0
langchain-core>=0.1.10
langchain-google-genai>=0.1.0
langgraph>=0.0.40
python-dotenv>=1.0.0
google-generativeai>=0.3.2
argparse>=1.4.0
colorama>=0.4.6
rich>=13.0.0 